refactor: package load test report scaffold
The load-test report scaffold is the base for query and Prometheus logic added later in the stack. Keeping it as a top-level Python file makes future modules use generic import names and increases the chance of import clashes once the tool grows beyond the placeholder command.

Move the placeholder command into a src-layout package, expose a console script, and keep the current no-op behavior covered by the scaffold tests. This gives the dependent PRs a stable package namespace without mixing query or report logic into this layer.
