+++ load-tests/docs/scripts/load_test_report/src/load_test_report/__main__.py
@@ -1,0 +1,3 @@
+from .cli import main
+
+main()
